Misc tweaks and code clean up
* Imagine: Add IO::writeVector() API to allow for writing out multiple buffers in a single syscall using writev()
* Imagine: Optimize PipeMessagePort::sendWithExtraData() using writeVector() API
* Imagine: Split out PipeMessages class from PipeMessagePort and add MessageReplyMode enum to replace bool parameters
* Imagine: Allow RendererTask::run() to append extra data after the CommandMessage
* Imagine: Optimize large data case in RendererTask::write() to avoid blocking on Buffer::map()
* Imagine: Add support for OpenGL Vertex Array Objects to reduce drawing overhead when supported
* Imagine: Make RendererTask in Gfx::Buffer immutable to simplify code
* Imagine: Add vertex array helper classes and update all templated drawing functions to use them
* Imagine: Move common quad index buffer from ViewManager to Renderer and add QuadVertexArray class that implicitly sets it
* Imagine: Pass ViewAttachParams to MenuItem objects so the constructors have access to the default fonts and renderer
* Imagine: Simplify MenuItem constructors using Config class
* Imagine: Combine all generic math headers into single math.hh
* Imagine: Abort program on a high severity OpenGL error in debug builds
* Imagine: Add support for running MessagePort without an event loop in blocking mode
* Imagine: Update GLTask to use MessagePort without an event loop to minimize overhead
* EmuFramework: Update Precise Frame Pacing option to allow for applying presentation time even when using the full target frame rate
* EmuFramework: Remove unneeded Image Buffers option
* EmuFramework: Fix crash when selecting yes in promptSystemReloadDueToSetOption()

+++ EmuFramework/src/shared/mednafen-emuex/ArchiveVFS.cc
@@ -45,7 +45,7 @@ Stream* ArchiveVFS::open(const std::string &path, const uint32 mode, const int d
 		if(name == filename)
 		{
 			auto stream = std::make_unique<MemoryStream>(entry.size(), true);
-			if(entry.read(stream->map(), entry.size()) != entry.size())
+			if(entry.read(stream->map(), entry.size()) != ssize_t(entry.size()))
 			{
 				throw MDFN_Error(0, "Error reading archive file:\n%s", filename.c_str());
 			}
